same as telnet but with fix login
first client will ask the server to connect. if server permits then client will start the connection and if not connection will be closed.
after server accepts the connection request, server sends confirmation to client and also asks for user name and password.
in this code user name and password is fixed by server(not of the server system user's credential).in my case it is user name = root and password = redhat.
after client enters these, client can execute commands on server and have out put on it's terminal.